Moved files to subpackage & added curly
