+++ jgrapht-core/src/main/java/org/jgrapht/alg/matching/blossom_v/DualUpdater.java
@@ -15,12 +15,12 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.matching;
+package org.jgrapht.alg.matching.blossom_v;
 
 import org.jgrapht.util.FibonacciHeap;
 
 import static java.lang.Math.abs;
-import static org.jgrapht.alg.matching.KolmogorovMinimumWeightPerfectMatching.*;
+import static org.jgrapht.alg.matching.blossom_v.KolmogorovMinimumWeightPerfectMatching.*;
 
 /**
  * This class is used by {@link KolmogorovMinimumWeightPerfectMatching} to perform dual updates, thus creating
