+++ jgrapht-core/src/main/java/org/jgrapht/alg/matching/blossom_v/Edge.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.matching;
+package org.jgrapht.alg.matching.blossom_v;
 
 import org.jgrapht.util.FibonacciHeapNode;
