+++ jgrapht-core/src/test/java/org/jgrapht/alg/matching/blossom_v/BlossomVDebugger.java
@@ -15,9 +15,8 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.matching;
+package org.jgrapht.alg.matching.blossom_v;
 
-import org.jgrapht.alg.matching.*;
 import org.jgrapht.util.FibonacciHeap;
 
 import java.util.HashSet;
