+++ jgrapht-core/src/test/java/org/jgrapht/alg/matching/blossom_v/PrimalUpdaterTest.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.matching;
+package org.jgrapht.alg.matching.blossom_v;
 
 import org.jgrapht.Graph;
 import org.jgrapht.Graphs;
@@ -28,9 +28,9 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 
-import static org.jgrapht.alg.matching.Initializer.InitializationType.NONE;
-import static org.jgrapht.alg.matching.KolmogorovMinimumWeightPerfectMatching.EPS;
-import static org.jgrapht.alg.matching.Node.Label.INFINITY;
+import static org.jgrapht.alg.matching.blossom_v.Initializer.InitializationType.NONE;
+import static org.jgrapht.alg.matching.blossom_v.KolmogorovMinimumWeightPerfectMatching.EPS;
+import static org.jgrapht.alg.matching.blossom_v.Node.Label.INFINITY;
 import static org.junit.Assert.*;
 
 public class PrimalUpdaterTest {
