+++ jgrapht-core/src/test/java/org/jgrapht/alg/matching/blossom_v/TreeEdgeTest.java
@@ -15,10 +15,11 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.matching;
+package org.jgrapht.alg.matching.blossom_v;
 
 import org.jgrapht.Graph;
 import org.jgrapht.Graphs;
+import org.jgrapht.alg.matching.blossom_v.*;
 import org.jgrapht.graph.DefaultUndirectedWeightedGraph;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.junit.Test;
@@ -26,7 +27,7 @@ import org.junit.Test;
 import java.util.Collections;
 import java.util.HashSet;
 
-import static org.jgrapht.alg.matching.Initializer.InitializationType.NONE;
+import static org.jgrapht.alg.matching.blossom_v.Initializer.InitializationType.NONE;
 import static org.junit.Assert.*;
 
 public class TreeEdgeTest {
